% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartVals.R
\name{getStartVals}
\alias{getStartVals}
\title{Obtaining starting values for Optim}
\usage{
getStartVals(y, covs, startVals, mar.mod)
}
\arguments{
\item{A}{vector of response, a matrix of covariates, start values if exist, type of Marginalized Distribution}
}
\value{
A vector of start values
}
\description{
The function creates starting values corresponding to the Marginalized Distributions for feed into Optim function
}
