% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePD.R
\name{makePD}
\alias{makePD}
\title{Positive Definite (PD) approximation of a Matrix}
\usage{
makePD(mat)
}
\arguments{
\item{mat}{Input a matrix}
}
\value{
A modified PD matrix
}
\description{
The function uses Eigen values and Eigen vectors to create a PD matrix
}
