% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAZIC.L1.R
\name{MAZIC.L1}
\alias{MAZIC.L1}
\title{Lasso estimates for MAZIC}
\usage{
MAZIC.L1(
  data,
  formula,
  startVals = TRUE,
  opt.method = c("Nelder-Mead", "BFGS"),
  mod = c("marzip", "marzinb", "marzigp"),
  verbose = TRUE
)
}
\arguments{
\item{The}{dataframe, the formula of the model, option for using starting values, optimization method,
names of the distribution, and verbose option}
}
\value{
The lasso estimates
}
\description{
The function returns the lasso estimates of each covariate
}
