% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAZIC.MLE.R
\name{MAZIC.MLE}
\alias{MAZIC.MLE}
\title{Maximum Likelihood estimates}
\usage{
MAZIC.MLE(
  y,
  x,
  startVals,
  opt.method = c("Nelder-Mead", "BFGS"),
  mod = c("marzip", "marzinb", "marzigp"),
  verbose = TRUE
)
}
\arguments{
\item{A}{vector of response, a matrix of covariates, starting values, optimization method,
names of the distribution, and verbose option}
}
\value{
The maximum likelihood estimates
}
\description{
The function returns the maximum likelihood estimates of each covariate
}
