% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAZIC.ridge.R
\name{MAZIC.ridge}
\alias{MAZIC.ridge}
\title{Ridge Regression estimates}
\usage{
MAZIC.ridge(
  y,
  x,
  startVals,
  opt.method = c("Nelder-Mead", "BFGS"),
  mod = c("marzip", "marzinb", "marzigp"),
  lambda,
  verbose = TRUE
)
}
\arguments{
\item{A}{vector of response, a matrix of covariates, starting values, optimization method,
names of the distribution, and verbose option}
}
\value{
The ridge estimates
}
\description{
The function returns the ridge estimates of each covariate
}
