% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgMetaData.R
\name{orgMetadata}
\alias{orgMetadata}
\title{Organize Design Matrix}
\usage{
orgMetadata(data, formula)
}
\arguments{
\item{mat}{A dataframe and the formula of the model}
}
\value{
An organized design matrix
}
\description{
The function properly organize design matrix for input to MAZIC L1 function
}
