% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.MarZip.R
\name{reg.marzip}
\alias{reg.marzip}
\title{Log Likelihood for Marginalized ZIP}
\usage{
reg.marzip(y, X, par, lambda = 0)
}
\arguments{
\item{A}{vector of response, a matrix of covariates, names of the parameter,
and tuning parameter for Ridge penalty term (default is 0)}
}
\value{
Log Likelihood of Marginalized ZIP
}
\description{
The function returns the log likelihood of a Marginalized ZIP
}
